Moved custom user model into its own app
